working

+++ ugbase/lib_algebra/operator/preconditioner/amg/amg.h
@@ -84,7 +84,7 @@ public:
 protected:
 //  functions
 	virtual void create_AMG_level(matrix_type &AH, SparseMatrix<double> &R, const matrix_type &A,
-							SparseMatrix<double> &P, int level);
+							SparseMatrix<double> &P, size_t level);
 
 // data
 	double eps_truncation_of_interpolation;	///< parameter used for truncation of interpolation
