+++ ugbase/lib_algebra/operator/preconditioner/amg/famg.h
@@ -91,15 +91,15 @@ public:
 	void set_theta(double theta) { m_theta = theta; }
 
 	void set_testvector_zero_at_dirichlet(bool bZeroAtDirichlet) { m_bTestvectorZeroAtDirichlet = bZeroAtDirichlet; }
-	void set_testvector_damps(int testvectordamps) { m_iTestvectorDamps = testvectordamps; }
+	void set_testvector_damps(size_t testvectordamps) { m_iTestvectorDamps = testvectordamps; }
 
 private:
 //  functions
 	virtual void create_AMG_level(matrix_type &AH, SparseMatrix<double> &R, const matrix_type &A,
-							SparseMatrix<double> &P, int level);
+							SparseMatrix<double> &P, size_t level);
 
 	void c_create_AMG_level(matrix_type &AH, SparseMatrix<double> &R, const matrix_type &A,
-			SparseMatrix<double> &P, int level);
+			SparseMatrix<double> &P, size_t level);
 
 private:
 // data
@@ -108,7 +108,7 @@ private:
 	double m_delta;								///< "Interpolation quality" F may not be worse than this (F < m_delta)
 	double m_theta;								///< clip all interpolations with m_theta * F > min F.
 
-	int m_iTestvectorDamps;
+	size_t m_iTestvectorDamps;
 	bool m_bTestvectorZeroAtDirichlet;
 
 };
